test : refactor SseControllerTest move package
